Game of Rock, Paper, Scissors, Lizard, Spock
The version in the last commit was not working properly past the first round of the game. The recurse of the gamePlay function was not properly coded. The return statement was missing in line116. The current version works from start to finish and tracks the score correctly regardless of the number of rounds players choose to play. My instructor critisized the function structure in this version, and I'm going to rewrite the code to make each function perform only one task.