Game of Rock, Paper, Scissors, Lizard, Spock
This is the rewritten code that my instructor accepted, but I still have a lot of doubts and questions about. I don't like the while loop for repeat rounds. I wanted to use recursion, and it seems as if I have to redo the main function to achieve this result.